+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Address deprecation warnings for create_function with PHP 7 (#193, thanks @chrisdeeming)
 * Add GrantedPaymentInstrumentUpdate webhook support
 * Allow paypal => options params to be passed in camel case
+* Add ability to create a transaction from a shared nonce
 
 ## 3.25.0
 * Add `submitForSettlement` option to `Subscription::retryCharge`